Update Z Select Trigger
Changed Z Select Trigger from Shift+Z to W+Q. Definitely won't work well for QWERTY users, but it's '+, for Dvorak users. I can create a QWERTY version that's mirrored to use the punctuation marks on the right hand side and the arrow keys on the left if anybody would like that.